Extract bookmark system as store
