+++ src/components/sidebar/tabs/nodeLibrary/NodeTreeFolder.vue
@@ -17,12 +17,12 @@
 
 <script setup lang="ts">
 import { ComfyNodeDefImpl } from '@/stores/nodeDefStore'
-import { useSettingStore } from '@/stores/settingStore'
+import { useNodeBookmarkStore } from '@/stores/nodeBookmarkStore'
 import type { CanvasDragAndDropData } from '@/types/litegraphTypes'
 import { dropTargetForElements } from '@atlaskit/pragmatic-drag-and-drop/element/adapter'
 import Badge from 'primevue/badge'
 import type { TreeNode } from 'primevue/treenode'
-import { computed, onMounted, onUnmounted, ref } from 'vue'
+import { onMounted, onUnmounted, ref } from 'vue'
 
 const props = defineProps<{
   node: TreeNode
@@ -33,11 +33,8 @@ const emit = defineEmits<{
   (e: 'itemDropped', node: TreeNode): void
 }>()
 
-const settingStore = useSettingStore()
-// Bookmarks are in format of category/display_name. e.g. "comfy/conditioning/CLIPTextEncode"
-const bookmarks = computed(() =>
-  settingStore.get('Comfy.NodeLibrary.Bookmarks')
-)
+const nodeBookmarkStore = useNodeBookmarkStore()
+
 const addNodeToBookmarkFolder = (node: ComfyNodeDefImpl) => {
   if (!props.node.data) {
     console.error('Bookmark folder does not have data!')
@@ -45,11 +42,7 @@ const addNodeToBookmarkFolder = (node: ComfyNodeDefImpl) => {
   }
   const folderNodeDef = props.node.data as ComfyNodeDefImpl
   const nodePath = folderNodeDef.category + '/' + node.display_name
-  if (bookmarks.value.includes(nodePath)) return
-  settingStore.set('Comfy.NodeLibrary.Bookmarks', [
-    ...bookmarks.value,
-    nodePath
-  ])
+  nodeBookmarkStore.addBookmark(nodePath)
 }
 
 const container = ref<HTMLElement | null>(null)
